Spoof User-Agent when talking to Google sites
Some Google sites do User-Agent sniffing and cripple the experience
for browsers that are not in their white lists. This includes not
delivering the nice fade-in effect for Google's main page, and HTML5
videos on Youtube. This change makes WebKitGTK+ fake being Chrome
through the User-Agent field whenever the "site specific quirks"
setting is enabled.